# api/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: "UA Consultants"
host: "13.211.139.75:3800"
basePath: "/"
...
schemes:
  - "http"
paths:
...
  /job/photo/upload:
    post:
      tags:
        - "job"
      summary: "Upload a job photo"
      description: ""
      operationId: "jobphotoupload"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/UploadJobPhoto"
      responses:
        200:
          description: "Photo uploaded successfully"
          schema:
            $ref: "#/definitions/UploadSuccess"
        400:
          description: "Photo upload failed"
          schema:
            $ref: "#/definitions/ResponseFail"
  /job/register:
    post:
      tags:
        - "job"
      summary: "Register a job"
      description: ""
      operationId: "jobregister"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/RegisterJob"
      responses:
        200:
          description: "Job registered successfully"
          schema:
            $ref: "#/definitions/ResponseSuccess"
        400:
          description: "Job failed to register"
          schema:
            $ref: "#/definitions/ResponseFail"
  /job/view:
    get:
      tags:
        - "job"
      summary: "View a job by id"
      description: ""
      operationId: "jobview"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/ViewJob"
      responses:
        200:
          description: "View job success"
          schema:
            $ref: "#/definitions/Projects"
        400:
          description: "View job error"
          schema:
            $ref: "#/definitions/ResponseFail"
...
definitions:
...
  ResponseSuccess:
    type: "object"
    properties:
      status:
        type: "string"
        example: "success"
  ResponseFail:
    type: "object"
    properties:
      error:
        type: "string"
      status:
        type: "string"
        example: "failure"
...
  Projects:
    type: "object"
    properties:
      project_name:
        type: "string"
        example: "Skirting board installation"
      date_created:
        type: "integer"
        example: 1568267256799
      description:
        type: "string"
      business_id:
        type: "string"
      business_name:
        type: "string"
        example: "Paul's skirtings"
      location:
        type: "string"
        example: "Newtown, NSW"
      current_bid:
        type: "integer"
      bid_amount:
        type: "integer"
      project_photos:
        type: "array"
        items:
          $ref: '#/definitions/ProjectPhoto'
...
  ProjectPhoto:
    type: "object"
    properties:
      image:
        type: "string"
        example: "https://images.unsplash.com/photo-1556020685-ae41abfc9365?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=400&q=60"
      title:
        type: "string"
...
  RegisterJob:
    type: "object"
    properties:
      session_token:
        type: "string"
      project_title:
        type: "string"
      project_description:
        type: "string"
      project_category:
        type: "string"
      project_type:
        type: "string"
      project_premium:
        type: "string"
      project_location:
        type: "string"
      project_photos:
        type: "array"
        items:
          $ref: '#/definitions/ProjectPhoto'
  UploadJobPhoto:
    type: "object"
    properties:
      session_token:
        type: "string"
      image_byte_array:
        type: "string"
  UploadSuccess:
    type: "object"
    properties:
      url:
        type: "string"
...
  ViewJob:
    type: "object"
    properties:
      job_id:
        type: "string"
      session_token:
        type: "string"
...